pyproject.toml
[project]
name = "ex-gpt-stt-uv"
version = "0.1.0"
description = "Enhanced faster-whisper with Korean STT post-processing and meeting minutes generation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faster-whisper>=1.2.0",
    "requests>=2.32.5",
    "psutil>=5.9.0",
    "nvidia-cudnn-cu12>=9.13.0.50",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]
gpu = [
    "nvidia-cudnn-cu12>=9.0.0",
]
benchmark = [
    "transformers>=4.25.0",
    "jiwer>=3.0.0",
    "datasets>=2.14.0",
    "memory-profiler>=0.61.0",
    "pytubefix>=6.0.0",
]
diarization = [
    "pyannote.audio>=3.1.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
]
